code/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
bedrock-agentcore>=0.1.5
bedrock-agentcore-starter-toolkit>=0.1.0
langgraph>=0.2.0
langchain-aws>=0.1.0
langchain-core>=0.3.0
streamlit>=1.28.0
PyYAML>=6.0
urllib3>=1.26.0
